% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cb_ctry}
\alias{cb_ctry}
\title{Country codes}
\format{A tibble:
\describe{
\item{country_name}{Straightforward name for country}
\item{ctry_sh}{Three-letter, lower case short code for country}
\item{iso3166_1a3}{The ISO 3166 code for the country. This is the three-letter code (alpha-3).
For further details, see \url{https://www.iso.org/iso-3166-country-codes.html}}
\item{iso3166_snamelc}{The ISO 3166 English short name}
\item{oegm_ctry_name}{Oxford Economics global model country name}
\item{oegm_ctry_twoletter}{Oxford Economics global model two-letter code}
\item{oegc_ctry_name}{Oxford Economics Global Cities country name}
\item{oegc_loc_code}{Oxford Economics Global Cities country location code}
\item{oegm_region_primary}{Oxford Economics global model primary region}
\item{oegts_country_name}{Oxford Economics Global Travel Service country name}
\item{oegts_region_midlev}{Oxford Economics Global Travel Service mid-level region}
\item{oegts_region_highlev}{Oxford Economics Global Travel Service high-level region. These include: Europe, Africa and Middle East, Americas, Asia Pacific, and NA.}
\item{oegts_origin_code}{Oxford Economics Global Travel Service origin code}
\item{note}{Note}}}
\usage{
cb_ctry
}
\description{
Country names and codes across sources.
}
\keyword{datasets}
