% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cb_var_oegc}
\alias{cb_var_oegc}
\title{Variable codes for Global Cities data}
\format{A tibble:
\describe{
\item{var}{Short, lower-case variable code (no special chracters)}
\item{oe_var}{Variable code in database (includes special characters such as $)}
\item{sh_description}{Short description of indicator}
\item{oe_description}{Description of indicator in database}}}
\usage{
cb_var_oegc
}
\description{
Short, lower-case variable codes to use with Global Cities data.
}
\keyword{datasets}
