% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{city_code_book}
\alias{city_code_book}
\title{City codes}
\format{A tibble:
\describe{
\item{mkt_name}{Straightforward name for city}
\item{mkt_sh}{Five-letter code for market}
\item{country_name}{Straightforward name for country. Useful to join with ctry_code_book}
\item{iso3166_1a3}{The ISO 3166 code for the country. This is the three-letter code (alpha-3).
For further details, see \url{https://www.iso.org/iso-3166-country-codes.html}}
\item{oe_globalcity_name}{Oxford Economics Global Cities city name}
\item{oegc_loc_code}{Oxford Economics Global Cities location code}
\item{oegc_ctry_name}{Oxford Economics Global Cities country name}
\item{str_text}{STR market name}
\item{mkt_geocode_friendly}{Market name that works well in geocoding scripts, such as Google API}
\item{oe_gct_name}{Oxford Economics Global City Travel city name}
\item{oe_gct_location_code}{Oxford Economics Global City Travel location code}}}
\usage{
city_code_book
}
\description{
City names and codes across sources.
}
\keyword{datasets}
