% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctry_code_book_oegtc}
\alias{ctry_code_book_oegtc}
\title{Country codes for Global City Travel data}
\format{A tibble:
\describe{
\item{orig_ctry_code}{Three-letter upper-case codes used in variables in GCT}
\item{country_name}{Straightforward name for country. Useful to join with ctry_code_book}
\item{oegtc_country_name}{country name used in GCT descriptions}
\item{type}{Type of geography (global, region, country)}
\item{oegtc_region_midlev}{Mid-level regions used in GCT}
\item{oegtc_region_highlev}{High-level regions used in GCT}}}
\usage{
ctry_code_book_oegtc
}
\description{
The origin country codes and regions used in Global City Travel
}
\keyword{datasets}
