% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_code_book_oegct}
\alias{var_code_book_oegct}
\title{Variable codes for Global City Travel data}
\format{A tibble:
\describe{
\item{var}{Short, lower-case variable code (no special chracters)}
\item{oe_var}{Variable code in database (includes special characters such as $)}
\item{sh_description}{Short description of indicator}
\item{oe_description}{Description of indicator in database}}}
\usage{
var_code_book_oegct
}
\description{
Short, lower-case variable codes to use with Global City Travel data.
}
\keyword{datasets}
